refactor: merge organism and page
