+++ src/components/pages/Work.scss
